bring development up to speed with master

+++ runtime/linux.tcl
@@ -1,4 +1,4 @@
-set VROOT_MASTER "docker.io/imunes/vroot:base"
+set VROOT_MASTER "imunes/vroot:base"
 
 #****f* linux.tcl/writeDataToNodeFile
 # NAME
@@ -193,8 +193,12 @@ proc spawnShell { node cmd } {
 # RESULT
 #   * exp_list -- experiment id list
 #****
-proc fetchRunningExperiments {} {}
-    # FIXME: make this work in Linux
+proc fetchRunningExperiments {} {
+    catch {exec himage -l | cut -d " " -f 1} exp_list
+    set exp_list [split $exp_list "
+"]
+    return $exp_list
+}
 
 #****f* linux.tcl/allSnapshotsAvailable
 # NAME
@@ -256,7 +260,6 @@ proc createExperimentContainer {} {}
 proc loadKernelModules {} {
     global all_modules_list
 
-    # FIXME: prepareSystem isn't the same on Linux
     foreach module $all_modules_list {
         if {[info procs $module.prepareSystem] == "$module.prepareSystem"} {
             $module.prepareSystem
@@ -350,20 +353,34 @@ proc createLinkBetween { lnode1 lnode2 ifname1 ifname2 } {
     }
     set ether2 [getIfcMACaddr $lnode2 $ifname2]
 
+    if {[nodeType $lnode1] == "rj45"} {
+        set lname1 [getNodeName $lnode1]
+    } else {
+        set lname1 $lnode1
+    }
+
+    if {[nodeType $lnode2] == "rj45"} {
+        set lname2 [getNodeName $lnode2]
+    } else {
+        set lname2 $lnode2
+    }
+
     if { [[typemodel $lnode1].virtlayer] == "NETGRAPH" } {
         if { [[typemodel $lnode2].virtlayer] == "NETGRAPH" } {
-            exec ovs-vsctl add-port $eid.$lnode1 $eid.$lnode1.$ifname1 -- \
-                set interface $eid.$lnode1.$ifname1 type=patch options:peer=$eid.$lnode2.$ifname2
-            exec ovs-vsctl add-port $eid.$lnode2 $eid.$lnode2.$ifname2 -- \
-                set interface $eid.$lnode2.$ifname2 type=patch options:peer=$eid.$lnode1.$ifname1
+            # generate interface names
+            set hostIfc1 "$eid.$lname1.$ifname1"
+            set hostIfc2 "$eid.$lname2.$ifname2"
+            # create veth pair
+            catch {exec ip link add name "$hostIfc1" type veth peer name "$hostIfc2"}
+            # add veth interfaces to bridges
+            exec ovs-vsctl add-port $eid.$lname1 $hostIfc1
+            exec ovs-vsctl add-port $eid.$lname2 $hostIfc2
+            # set bridge interfaces up
+            exec ip link set dev $hostIfc1 up
+            exec ip link set dev $hostIfc2 up
         }
         if { [[typemodel $lnode2].virtlayer] == "VIMAGE" } {
-            if {[nodeType $lnode1] == "rj45"} {
-                set extIfcName [getNodeName $lnode1]
-                addNodeIfcToBridge $extIfcName $ifname2 $lnode2 $ether2
-            } else {
-                addNodeIfcToBridge $lnode1 $ifname2 $lnode2 $ether2
-            }
+            addNodeIfcToBridge $lname1 $ifname1 $lnode2 $ifname2 $ether2
         }
     } elseif { [[typemodel $lnode1].virtlayer] == "VIMAGE" } {
         if  { [[typemodel $lnode2].virtlayer] == "VIMAGE" } {
@@ -388,17 +405,37 @@ proc createLinkBetween { lnode1 lnode2 ifname1 ifname2 } {
             exec rm -f "/var/run/netns/$lnode2Ns"
         }
         if { [[typemodel $lnode2].virtlayer] == "NETGRAPH" } {
-            if {[nodeType $lnode2] == "rj45"} {
-                set extIfcName [getNodeName $lnode2]
-                addNodeIfcToBridge $extIfcName $ifname1 $lnode1 $ether1
-            } else {
-                addNodeIfcToBridge $lnode2 $ifname1 $lnode1 $ether1
-            }
+            addNodeIfcToBridge $lname2 $ifname2 $lnode1 $ifname1 $ether1
         }
     }
 }
 
-proc configureLinkBetween { lnode1 lnode2 ifname1 ifname2 link } {}
+proc configureLinkBetween { lnode1 lnode2 ifname1 ifname2 link } {
+    upvar 0 ::cf::[set ::curcfg]::eid eid
+
+    # FIXME: merge this with execSet* commands
+    execSetLinkParams $eid $link
+
+    # if {[nodeType $lnode1] != "rj45" && [nodeType $lnode2] != "rj45"} {
+    #     set qdisc [getIfcQDisc $lnode1 $ifname1]
+    #     if {$qdisc ne "FIFO"} {
+    #         execSetIfcQDisc $eid $lnode1 $ifname2 $qdisc
+    #     }
+
+    #     set qdisc [getIfcQDisc $lnode2 $ifname2]
+    #     if {$qdisc ne "FIFO"} {
+    #         execSetIfcQDisc $eid $lnode2 $ifname2 $qdisc
+    #     }
+    #     set qdrop [getIfcQDrop $node $ifc]
+    #     if {$qdrop ne "drop-tail"} {
+    #         execSetIfcQDrop $eid $node $ifc $qdrop
+    #     }
+    #     set qlen [getIfcQLen $node $ifc]
+    #     if {$qlen ne 50} {
+    #         execSetIfcQLen $eid $node $ifc $qlen
+    #     }
+    # }
+}
 
 #****f* linux.tcl/startIfcsNode
 # NAME
@@ -453,6 +490,9 @@ proc runConfOnNode { node } {
     set node_dir [getVrootDir]/$eid/$node
     set node_id "$eid.$node"
 
+    catch {exec docker exec $node_id umount /etc/resolv.conf}
+    catch {exec docker exec $node_id umount /etc/hosts}
+
     if { [getCustomEnabled $node] == true } {
         set selected [getCustomConfigSelected $node]
 
@@ -465,22 +505,24 @@ proc runConfOnNode { node } {
         set confFile "boot.conf"
     }
 
-    catch {exec docker inspect --format '{{.Id}}' $node_id} id
     writeDataToFile $node_dir/$confFile [join $bootcfg "\n"]
     exec docker exec -i $node_id sh -c "cat > $confFile" < $node_dir/$confFile
     exec docker exec $node_id $bootcmd $confFile >& $node_dir/out.log &
     exec docker exec -i $node_id sh -c "cat > out.log" < $node_dir/out.log
 
     foreach ifc [allIfcList $node] {
-        if {[getIfcOperState $node $ifc] == "down"} {
-            exec docker exec $node_id ip link set dev $ifc down
+        # FIXME: should also work for loopback
+        if {$ifc != "lo0"} {
+            if {[getIfcOperState $node $ifc] == "down"} {
+                exec docker exec $node_id ip link set dev $ifc down
+            }
         }
     }
 }
 
 proc destroyLinkBetween { eid lnode1 lnode2 } {
-    set lname [linkByPeers $lnode1 $lnode2]
-    catch {exec ovs-vsctl del-br $eid.$lname}
+    set ifname1 [ifcByLogicalPeer $lnode1 $lnode2]
+    catch {exec ip link del dev $eid.$lnode1.$ifname1}
 }
 
 #****f* linux.tcl/removeNodeIfcIPaddrs
@@ -497,7 +539,10 @@ proc destroyLinkBetween { eid lnode1 lnode2 } {
 proc removeNodeIfcIPaddrs { eid node } {
     set node_id "$eid.$node"
     foreach ifc [allIfcList $node] {
-        catch "exec docker exec $node_id ip addr flush dev $ifc"
+        # FIXME: make this work for loopback
+        if {$ifc != "lo0"} {
+            catch "exec docker exec $node_id ip addr flush dev $ifc"
+        }
     }
 }
 
@@ -592,6 +637,8 @@ proc l2node.destroy { eid node } {
 proc enableIPforwarding { eid node } {
     pipesExec "docker exec $eid\.$node sysctl net.ipv6.conf.all.forwarding=1" "hold"
     pipesExec "docker exec $eid\.$node sysctl net.ipv4.conf.all.forwarding=1" "hold"
+    pipesExec "docker exec $eid\.$node sysctl net.ipv4.conf.default.rp_filter=0" "hold"
+    pipesExec "docker exec $eid\.$node sysctl net.ipv4.conf.all.rp_filter=0" "hold"
 }
 
 #****f* linux.tcl/configDefaultLoIfc
@@ -730,7 +777,7 @@ proc getNodeNamespace { node } {
     return $ns
 }
 
-proc addNodeIfcToBridge { bridge ifc node mac } {
+proc addNodeIfcToBridge { bridge brifc node ifc mac } {
     upvar 0 ::cf::[set ::curcfg]::eid eid
 
     set nodeNs [createNetNs $node]
@@ -738,13 +785,14 @@ proc addNodeIfcToBridge { bridge ifc node mac } {
     catch "exec ovs-vsctl add-br $eid.$bridge"
 
     # generate interface names
-    set hostIfc "v${ifc}pl${nodeNs}"
-    set guestIfc "v${ifc}pg${nodeNs}"
+    set hostIfc "$eid.$bridge.$brifc"
+    set guestIfc "$eid.$node.$ifc"
 
     # create veth pair
     exec ip link add name "$hostIfc" type veth peer name "$guestIfc"
     # add host side of veth pair to bridge
     exec ovs-vsctl add-port "$eid.$bridge" "$hostIfc"
+
     exec ip link set "$hostIfc" up
 
     # move guest side of veth pair to node namespace
@@ -788,4 +836,210 @@ proc setIfcNetNs { node oldIfc newIfc } {
     set nodeNs [getNodeNamespace $node]
     exec ip link set "$oldIfc" netns "$nodeNs"
     exec ip netns exec "$nodeNs" ip link set "$oldIfc" name "$newIfc"
+}
+
+#****f* linux.tcl/execSetIfcQDisc
+# NAME
+#   execSetIfcQDisc -- in exec mode set interface queuing discipline
+# SYNOPSIS
+#   execSetIfcQDisc $eid $node $ifc $qdisc
+# FUNCTION
+#   Sets the queuing discipline during the simulation.
+#   New queuing discipline is defined in qdisc parameter.
+#   Queueing discipline can be set to fifo, wfq or drr.
+# INPUTS
+#   eid -- experiment id
+#   node -- node id
+#   ifc -- interface name
+#   qdisc -- queuing discipline
+#****
+proc execSetIfcQDisc { eid node ifc qdisc } {
+    set target [linkByIfc $node $ifc]
+    set peers [linkPeers [lindex $target 0]]
+    set dir [lindex $target 1]
+    set lnode1 [lindex $peers 0]
+    set lnode2 [lindex $peers 1]
+    if { [nodeType $lnode2] == "pseudo" } {
+        set mirror_link [getLinkMirror [lindex $target 0]]
+        set lnode2 [lindex [linkPeers $mirror_link] 0]
+    }
+    switch -exact $qdisc {
+        FIFO { set qdisc fifo_fast }
+        WFQ { set qdisc wfq }
+        DRR { set qdisc drr }
+    }
+    exec docker exec $eid.$node tc qdisc add dev $ifc root $qdisc
+}
+
+#****f* linux.tcl/execSetLinkParams
+# NAME
+#   execSetLinkParams -- in exec mode set link parameters
+# SYNOPSIS
+#   execSetLinkParams $eid $link
+# FUNCTION
+#   Sets the link parameters during the simulation.
+#   All the parameters are set at the same time.
+# INPUTS
+#   eid -- experiment id
+#   link -- link id
+#****
+proc execSetLinkParams { eid link } {
+    global debug
+
+    set lnode1 [lindex [linkPeers $link] 0]
+    set lnode2 [lindex [linkPeers $link] 1]
+    set ifname1 [ifcByLogicalPeer $lnode1 $lnode2]
+    set ifname2 [ifcByLogicalPeer $lnode2 $lnode1]
+
+    set bandwidth [expr [getLinkBandwidth $link] + 0]
+    set delay [expr [getLinkDelay $link] + 0]
+    set ber [expr [getLinkBER $link] + 0]
+    set dup [expr [getLinkDup $link] + 0]
+
+    if {[nodeType $lnode1] == "rj45"} {
+        set lname1 [getNodeName $lnode1]
+    } else {
+        set lname1 $lnode1
+    }
+
+    if {[nodeType $lnode2] == "rj45"} {
+        set lname2 [getNodeName $lnode2]
+    } else {
+        set lname2 $lnode2
+    }
+
+    if { [[typemodel $lnode1].virtlayer] == "NETGRAPH" } {
+        catch {exec tc qdisc del dev $eid.$lname1.$ifname1 root}
+
+        set vdelay [expr $delay / 1000]
+        exec tc qdisc add dev $eid.$lname1.$ifname1 root \
+            handle 1: netem delay ${vdelay}ms
+
+        exec tc qdisc add dev $eid.$lname1.$ifname1 parent 1: \
+            handle 2: netem duplicate ${dup}%
+
+        set corrupt [expr (1 / double($ber)) * 100]
+        exec tc qdisc add dev $eid.$lname1.$ifname1 parent 2: \
+            handle 3: netem corrupt ${corrupt}%
+
+        if {$bandwidth > 0} {
+            exec tc qdisc add dev $eid.$lname1.$ifname1 parent 3: \
+                handle 4: tbf rate ${bandwidth}bit limit 10mb burst 1540
+        }
+    }
+
+    if { [[typemodel $lnode2].virtlayer] == "NETGRAPH" } {
+        catch {exec tc qdisc del dev $eid.$lname2.$ifname2 root}
+
+        set vdelay [expr $delay / 1000]
+        exec tc qdisc add dev $eid.$lname2.$ifname2 root \
+            handle 1: netem delay ${vdelay}ms
+
+        exec tc qdisc add dev $eid.$lname2.$ifname2 parent 1: \
+            handle 2: netem duplicate ${dup}%
+
+        set corrupt [expr (1 / double($ber)) * 100]
+        exec tc qdisc add dev $eid.$lname2.$ifname2 parent 2: \
+            handle 3: netem corrupt ${corrupt}%
+
+        if {$bandwidth > 0} {
+            exec tc qdisc add dev $eid.$lname2.$ifname2 parent 3: \
+                handle 4: tbf rate ${bandwidth}bit limit 10mb burst 1540
+        }
+    }
+
+    if { [[typemodel $lnode1].virtlayer] == "VIMAGE" } {
+        catch {exec docker exec $eid.$lnode1 tc qdisc del dev $ifname1 root}
+
+        set vdelay [expr $delay / 1000]
+        exec docker exec $eid.$lnode1 tc qdisc add dev $ifname1 root \
+            handle 1: netem delay ${vdelay}ms
+
+        exec docker exec $eid.$lnode1 tc qdisc add dev $ifname1 parent 1: \
+            handle 2: netem duplicate ${dup}%
+
+        set corrupt [expr (1 / double($ber)) * 100]
+        exec docker exec $eid.$lnode1 tc qdisc add dev $ifname1 parent 2: \
+            handle 3: netem corrupt ${corrupt}%
+
+        if {$bandwidth > 0} {
+            exec docker exec $eid.$lnode1 tc qdisc add dev $ifname1 parent 3: \
+                handle 4: tbf rate ${bandwidth}bit limit 10mb burst 1540
+        }
+
+    }
+
+    if { [[typemodel $lnode2].virtlayer] == "VIMAGE" } {
+        catch {exec docker exec $eid.$lnode2 tc qdisc del dev $ifname2 root}
+
+        set vdelay [expr $delay / 1000]
+        exec docker exec $eid.$lnode2 tc qdisc add dev $ifname2 root \
+            handle 1: netem delay ${vdelay}ms
+
+        exec docker exec $eid.$lnode2 tc qdisc add dev $ifname2 parent 1: \
+            handle 2: netem duplicate ${dup}%
+
+        set corrupt [expr (1 / double($ber)) * 100]
+        exec docker exec $eid.$lnode2 tc qdisc add dev $ifname2 parent 2: \
+            handle 3: netem corrupt ${corrupt}%
+
+        if {$bandwidth > 0} {
+            exec docker exec $eid.$lnode2 tc qdisc add dev $ifname2 parent 3: \
+                handle 4: tbf rate ${bandwidth}bit limit 10mb burst 1540
+        }
+    }
+}
+
+#****f* linux.tcl/startIPsecOnNode
+# NAME
+#   startIPsecOnNode -- start ipsec on node
+# SYNOPSIS
+#   startIPsecOnNode $eid $node
+# FUNCTION
+#   Starts strongswan ipsec daemons on the given node.
+#****
+proc startIPsecOnNode { eid node } {
+    catch {exec docker exec $eid\.$node ipsec start} err
+}
+
+proc ipsecFilesToNode { eid node local_cert ipsecret_file } {
+    set node_id "$eid\.$node"
+    set hostname [getNodeName $node]
+
+    if { $local_cert != "" } {
+	set trimmed_local_cert [lindex [split $local_cert /] end]
+	catch {exec hcp $local_cert $hostname@$eid:/etc/ipsec.d/certs/$trimmed_local_cert}
+    }
+
+    if { $ipsecret_file != "" } {
+	set fileId2 [open /tmp/imunes_$node_id\_ipsec.secrets w]
+	puts $fileId2 "# /etc/ipsec.secrets - strongSwan IPsec secrets file\n"
+	set trimmed_local_key [lindex [split $ipsecret_file /] end]
+	catch {exec hcp $ipsecret_file $hostname@$eid:/etc/ipsec.d/private/$trimmed_local_key}
+	puts $fileId2 ": RSA $trimmed_local_key"
+	close $fileId2
+    }
+
+    catch {exec hcp /tmp/imunes_$node_id\_ipsec.conf $hostname@$eid:/etc/ipsec.conf}
+    catch {exec hcp /tmp/imunes_$node_id\_ipsec.secrets $hostname@$eid:/etc/ipsec.secrets}
+}
+
+proc sshServiceStartCmds {} {
+    lappend cmds "dpkg-reconfigure openssh-server"
+    lappend cmds "service ssh start"
+    return $cmds
+}
+
+proc sshServiceStopCmds {} {
+    return "service ssh stop"
+}
+
+proc inetdServiceRestartCmds {} {
+    return "service openbsd-inetd restart"
+}
+
+proc moveFileFromNode { node path ext_path } {
+    upvar 0 ::cf::[set ::curcfg]::eid eid
+    catch {exec hcp [getNodeName $node]@$eid:$path $ext_path}
+    catch {exec docker exec $eid.$node rm -fr $path}
 }
